requirements_dev.txt
bumpversion==0.5.3
django-model-utils>=2.0
wheel==0.32.1
twine==1.12.1
django-extensions
